+++ com.unity.cinemachine/Runtime/Behaviours/CinemachineConfiner2D.cs
@@ -1,4 +1,5 @@
 #if CINEMACHINE_PHYSICS_2D
+
 using System;
 using System.Collections.Generic;
 using Cinemachine.Utility;