+++ com.unity.cinemachine/Runtime/Core/ConfinerOven.cs
@@ -1,6 +1,7 @@
+#if CINEMACHINE_PHYSICS_2D
+
 using System;
 using System.Collections.Generic;
-using System.Linq;
 using Cinemachine.Utility;
 using UnityEngine;
 
@@ -583,3 +584,4 @@ namespace Cinemachine
         }
     }
 }
+#endif